% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dist.R
\name{get_beta_parameters}
\alias{get_beta_parameters}
\title{Get beta distribution parameters}
\usage{
get_beta_parameters(population_avg, population_var)
}
\arguments{
\item{population_avg}{the population average, a length 1 numeric value
describing the average (mean) value of the population}

\item{population_var}{the population variance, a length 1 numeric value
describing the variance of the population}
}
\value{
a list object with numeric elements
}
\description{
Based on the population average and variance, calculate the alpha and beta
parameters that describe the beta distribution.
}
\details{
The beta distribution should be used when the distribution being predicting
is a value between 0 and 1.
}
\examples{
## Generate data following a beta distribution
library(MASS)

brain_standardized <- scale(Animals$brain)
beta_data <- (brain_standardized - min(brain_standardized)) / (max(brain_standardized) - min(brain_standardized))

beta_mean <- mean(beta_data)
beta_var <- var(beta_data)

## Using the function
beta_parameters <- get_beta_parameters(
 population_avg = beta_mean,
 population_var = beta_var
)
beta_parameters

}
